refactor(racingcar): 네이밍 변경, 패키지 변경
