+++ src/main/java/racingcar/view/InputValidation.java
@@ -1,4 +1,4 @@
-package racingcar.validation;
+package racingcar.view;
 
 
 public class InputValidation {
